moved logout function to header
